oke lanjut di laptop

+++ src/main/kotlin/com/erp/distribution/desgreenrestkt/data/source/entity/FCustomerEntity.kt
@@ -24,7 +24,7 @@ data class FCustomerEntity (
         initialValue = 912345668
     )
     @GeneratedValue(strategy = GenerationType.AUTO, generator="hibernate_sequence")
-    var id :Long =0L,
+    var id :Int =0,
 
     /*
 	 * JIKA COPY DARI TEMPAT LAIN: MAKA SEBAGAI LOG TRACK MENINGGALKAN SOURCE_ID = ID sumber asal dia dicopy
@@ -33,7 +33,7 @@ data class FCustomerEntity (
 	 * 2. 
 	 */
     @Column(name = "SOURCE_ID", length = 9)
-    var sourceID :Long =0L,
+    var sourceID :Int =0,
 
     @Column(name = "CUSTNO", length = 25)
     var custno :String ="",
